Fb 69 repo manager init

+++ repository_manager.py
@@ -3,6 +3,8 @@ import os
 import logging
 from typing import Optional
 
+from logging_interface import LoggingInterface
+
 
 class RepositoryUnsafeToDeleteException(Exception):
     def __init__(self, path):
@@ -34,7 +36,7 @@ class RepositoryPullException(Exception):
         super().__init__(f"Pulling repository failed.")
 
 
-class RepositoryManager:
+class RepositoryManager(LoggingInterface):
     """
     Class that allows for management of a git repository
     """
@@ -52,6 +54,7 @@ class RepositoryManager:
         :param repository_name: The name of the folder the repository exists or should be created in
         :param remote_url: The remote url the repository should be cloned from if needed
         """
+        super().__init__()
         self._safe_mode = False
         self._base_path = os.path.abspath(base_path)
         self._repo_name = repository_name
@@ -60,7 +63,6 @@ class RepositoryManager:
         else:
             self._path = self._base_path
         self._remote_url = remote_url
-        self._logger = logging.getLogger("RepositoryManager")
 
     def _exec_git_command(self, command: str) -> bool:
         """
@@ -68,7 +70,9 @@ class RepositoryManager:
         :param command: The command to execute
         :return: True if return code of command is 0
         """
-        return os.system(f"cd {self._path};{command}") == 0
+        command = f"cd {self._path};{command}"
+        self._logger.debug(f"Executing '{command}'")
+        return os.system(command) == 0
 
     def _dir_is_safe_to_delete(self) -> bool:
         """
@@ -83,7 +87,7 @@ class RepositoryManager:
 
     def set_safety(self, safe_mode: bool):
         """
-        By default, the manager can only delete repositories inside of its parent scripts working directory.\n
+        By default, the manager can only delete repositories in its parent scripts working directory.\n
         Setting this to 'False' allows the Manager to delete Folders everywhere on the disk.\n
         Beware you could delete your whole hard disk by messing this up.
         """
@@ -95,7 +99,7 @@ class RepositoryManager:
         :raises RepositoryUnsafeToDeleteException: When the repository folder is not safe to delete
         """
         if not self._dir_is_safe_to_delete():
-            raise RepositoryUnsafeToDeleteException()
+            raise RepositoryUnsafeToDeleteException(self._path)
 
         for root, dirs, files in os.walk(self._path, topdown=False):
             for name in files:
@@ -103,10 +107,11 @@ class RepositoryManager:
             for name in dirs:
                 os.rmdir(os.path.join(root, name))
 
-    def init_repository(self, force_reset: bool):
+    def init_repository(self, force_reset: bool = False, reclone_on_error: bool = True):
         """
         Checks if the repository exists and works. Clones the repository if it does not exist at the given path.
         :param force_reset: Whether repository should be deleted and re-clone
+        :param reclone_on_error: Whether repository should be recloned if an error occurs
         :raises RepositoryCloneException: When cloning the repository fails
         """
         repo_works = False
@@ -116,12 +121,16 @@ class RepositoryManager:
                 self.delete_folder()
         else:
             if os.path.isdir(self._path):
-                repo_clean = os.system(f"cd {self._path};git diff --quiet") == 0
+                repo_clean = self._exec_git_command("git diff --quiet")
                 repo_works = repo_clean
 
         if not repo_works:
+            if not reclone_on_error:
+                raise RepositoryStatusException
+
             self._logger.info(f"Repo doesn't exist or is broken, deleting dir"
                               f"and cloning repository from '{self._remote_url}'")
+
             if os.path.isdir(self._path):
                 self.delete_folder()
 
@@ -185,12 +194,13 @@ class RepositoryManager:
             raise RepositoryPullException()
         self._logger.info("Pulling from remote repository was successful")
 
-    def get_commit_hash(self) -> str:
+    def get_commit_hash(self, branch: str = "HEAD") -> str:
         """
         Gets the current git commit hash.
+        :param branch: The branch to get the commit hash for. defaults to HEAD
         :return: The current commit hash.
         """
-        return os.popen(f"cd {self._path};git rev-parse HEAD").read().strip("\n")
+        return os.popen(f"cd {self._path};git rev-parse {branch}").read().strip("\n")
 
     def get_branch_date(self) -> str:
         """
@@ -218,3 +228,17 @@ class RepositoryManager:
         if len(branch_list) != 1:
             raise RepositoryStatusException()
         return branch_list[0]
+
+    def get_remote_branch(self) -> str:
+        """
+        Gets the remote branch associated with the current local branch.
+        :return: The current remote branch.
+        """
+        branch_list = [x.strip().strip("*").strip()
+                       for x
+                       in os.popen(f"cd {self._path};git branch -vv").read().strip("\n").split("\n")
+                       if x.strip().startswith("*")]
+        if len(branch_list) != 1:
+            raise RepositoryStatusException()
+        data = [x for x in branch_list[0].split(" ") if x]
+        return data[2].split(": ")[0]
